requirements.txt
wheel==0.38.4
setuptools==65.5.0
gym==0.21.0
h5py==3.7.0
matplotlib==3.7.5
matplotlib-inline==0.1.3
mne==1.6.1
notebook==6.4.8
numpy==1.17.4
opencv-python==4.6.0.66
Pillow==9.0.1
PyYAML==6.0
scikit-learn==1.0.2
scipy==1.8.0
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.62.3
